app/requirements.txt
streamlit
tesserocr
torch
torchvision
pillow
transformers
datasets
tqdm
